pyproject.toml
[project]
name = "lanchainbuild"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.4.4",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "python-dotenv>=1.1.1",
]

[tool.uv.pip]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
